requirements.txt
asgiref==3.4.1
Django==3.2.6
django-countries==7.2.1
django-crispy-forms==1.12.0
django-froala-editor==3.2.2
django-phonenumber-field==5.2.0
phonenumbers==8.12.29
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
